Refactored to support PSR-4 autoloading support and file structure.
