+++ controllers/ControllerCommandFactory.php
@@ -1,7 +1,7 @@
 <?php
 /**
 * User Directory
-*   Copyright © 2008 Theodore R. Smith <theodore@phpexperts.pro>
+*   Copyright ï¿½ 2008 Theodore R. Smith <theodore@phpexperts.pro>
 * 
 * The following code is licensed under a modified BSD License.
 * All of the terms and conditions of the BSD License apply with one
@@ -14,6 +14,10 @@
 *
 * BSD License: http://www.opensource.org/licenses/bsd-license.php
 **/
+
+namespace PHPExperts\UserDirectory\Controllers;
+
+use Exception;
 
 class ControllerCommandFactory
 {
@@ -29,7 +33,8 @@ class ControllerCommandFactory
 		$controllers = array('UserController', 'SearchController', 'SecurityController');
 		foreach ($controllers as $c)
 		{
-			$controller = new $c;
+		    $className = 'PHPExperts\UserDirectory\Controllers\\' . $c;
+			$controller = new $className;
 			if (($output = $controller->execute($action)) !== false)
 			{
 				return $output;
