+++ lib/MyDB/MyDBI.php
@@ -1,0 +1,34 @@
+<?php
+/**
+ * User Directory
+ *   Copyright (c) 2008, 2012, 2019 Theodore R. Smith <theodore@phpexperts.pro>
+ *
+ * The following code is licensed under a modified BSD License.
+ * All of the terms and conditions of the BSD License apply with one
+ * exception:
+ *
+ * 1. Every one who has not been a registered student of the "PHPExperts
+ *    From Beginner To Pro" course (http://www.phpexperts.pro/) is forbidden
+ *    from modifing this code or using in an another project, either as a
+ *    deritvative work or stand-alone.
+ *
+ * BSD License: http://www.opensource.org/licenses/bsd-license.php
+ **/
+
+namespace PHPExperts\MyDB;
+
+interface MyDBI
+{
+    // Basic SQL functions
+    public function query($sql, array $params = null);
+    public function fetchArray();
+    public function fetchObject($objectType = '\stdClass');
+
+    // Transactional SQL functions
+    public function beginTransaction();
+    public function commit();
+    public function rollback();
+
+    // Write SQL functions
+    public function lastInsertId();
+}
