+++ managers/UserInfoStruct.php
@@ -1,7 +1,10 @@
 <?php
+
+namespace PHPExperts\UserDirectory\Managers;
+
 /**
 * User Directory
-*   Copyright © 2008 Theodore R. Smith <theodore@phpexperts.pro>
+*   Copyright ï¿½ 2008 Theodore R. Smith <theodore@phpexperts.pro>
 * 
 * The following code is licensed under a modified BSD License.
 * All of the terms and conditions of the BSD License apply with one
