+++ phpunit.xml
@@ -7,8 +7,8 @@
          forceCoversAnnotation="false"
          mapTestClassNameToCoveredClassName="true"
          processIsolation="false"
-         stopOnError="true"
-         stopOnFailure="true"
+         stopOnError="false"
+         stopOnFailure="false"
          stopOnIncomplete="false"
          stopOnSkipped="false"
          syntaxCheck="true"
