+++ tests/Mocks/MyDB_Engine_Mock.php
@@ -1,0 +1,63 @@
+<?php
+/**
+* User Directory
+*   Copyright (c) 2008, 2011, 2019 Theodore R. Smith <theodore@phpexperts.pro>
+* 
+* The following code is licensed under a modified BSD License.
+* All of the terms and conditions of the BSD License apply with one
+* exception:
+*
+* 1. Every one who has not been a registered student of the "PHPExperts
+*    From Beginner To Pro" course (http://www.phpexperts.pro/) is forbidden
+*    from modifing this code or using in an another project, either as a
+*    deritvative work or stand-alone.
+*
+* BSD License: http://www.opensource.org/licenses/bsd-license.php
+**/
+
+namespace Tests\PHPExperts\UserDirectory\Mocks;
+
+use Exception;
+use PHPExperts\MyDB\MyDBI;
+
+class MyDB_Engine_Mock implements MyDBI
+{
+	// Basic SQL functions
+	public function query($sql, array $params = null)
+	{
+		throw new Exception('Hit ' . __METHOD__);
+	}
+
+	public function fetchArray()
+	{
+		throw new Exception('Hit ' . __METHOD__);
+	}
+
+	public function fetchObject($objectType = '\stdClass')
+	{
+		throw new Exception('Hit ' . __METHOD__);
+	}
+
+	
+	// Transactional SQL functions
+	public function beginTransaction()
+	{
+		throw new Exception('Hit ' . __METHOD__);
+	}
+
+	public function commit()
+	{
+		throw new Exception('Hit ' . __METHOD__);
+	}
+
+	public function rollback()
+	{
+		throw new Exception('Hit ' . __METHOD__);
+	}
+	
+	// Write SQL functions
+	public function lastInsertId()
+	{
+		throw new Exception('Hit ' . __METHOD__);
+	}
+}
