+++ tests/MyDB/MyDBHelperFunctionsTest.php
@@ -16,9 +16,18 @@
  * BSD License: http://www.opensource.org/licenses/bsd-license.php
  **/
 
+namespace Tests\PHPExperts\MyDB;
+
+use function PHPExperts\MyDB\getDBHandler;
+use PHPExperts\MyDB\MyDBException;
+use PHPExperts\MyDB\MyDBI;
+use function PHPExperts\MyDB\queryDB;
+use PHPUnit\Framework\TestCase;
+use Tests\PHPExperts\UserDirectory\Mocks\MyDB_Engine_Mock;
+
 require_once __DIR__ . '/MyDatabaseTestSuite.php';
 
-class MyDBHelperFunctionsTest extends \PHPUnit\Framework\TestCase
+class MyDBHelperFunctionsTest extends TestCase
 {
     /**
      * Tests getDBHandler()
@@ -41,12 +50,12 @@ class MyDBHelperFunctionsTest extends \PHPUnit\Framework\TestCase
         // (change current directory to be able to find config path)
         chdir(dirname(__FILE__) . '/..');
         $pdo = getDBHandler();
-        $this->assertInstanceOf('MyDBI', $pdo, 'PDO object is not of type PDO');;
+        $this->assertInstanceOf(MyDBI::class, $pdo, 'PDO object is not of type PDO');;
 
         // Test with custom config
         $config = MyDatabaseTestSuite::getPDOConfig();
-        $new_pdo = getDBHandler($config);
-        $this->assertInstanceOf('MyDBI', $new_pdo, 'PDO object is not of type PDO');;
+        $new_pdo = getDBHandler($config, MyDB_Engine_Mock::class);
+        $this->assertInstanceOf(MyDBI::class, $new_pdo, 'PDO object is not of type PDO');;
     }
 
     /**
@@ -74,7 +83,7 @@ class MyDBHelperFunctionsTest extends \PHPUnit\Framework\TestCase
         $this->assertSame($username, $userInfo->username);
 
         // Test select w/ malformed SQL
-        $this->expectException('MyDBException');
-        $stmt = @queryDB('SELECT * FROM usersasdf WHERE username=?', array($username));
+        $this->expectException(MyDBException::class);
+        @queryDB('SELECT * FROM usersasdf WHERE username=?', array($username));
     }
 }
