+++ tests/bootstrap.inc.php
@@ -3,8 +3,7 @@
 // Use composer's autoload.
 require_once __DIR__ . '/../vendor/autoload.php';
 
-require_once __DIR__ . '/../lib/bootstrap.inc.php';
-
-require_once APP_PATH . '/lib/MyDB.inc.php';
-require_once APP_PATH . '/lib/UserInfoStruct.inc.php';
-require_once 'mocks.php';
+// Needed to resolve:
+// 1) Tests\PHPExperts\UserDirectory\SearchControllerTest::test__construct
+//    session_start(): Cannot start session when headers already sent
+ob_start();
