+++ views/browse.tpl.php
@@ -1,8 +1,8 @@
 <?php
 /**
 * User Directory
-*   Copyright © 2008 Theodore R. Smith <theodore@phpexperts.pro>
-* 
+*   Copyright (c) 2008, 2019 Theodore R. Smith <theodore@phpexperts.pro>
+*
 * The following code is licensed under a modified BSD License.
 * All of the terms and conditions of the BSD License apply with one
 * exception:
