+++ views/header.tpl.php
@@ -1,7 +1,7 @@
 <?php
 /**
 * User Directory
-*   Copyright © 2008 Theodore R. Smith <theodore@phpexperts.pro>
+*   Copyright (c) 2008, 2019 Theodore R. Smith <theodore@phpexperts.pro>
 * 
 * The following code is licensed under a modified BSD License.
 * All of the terms and conditions of the BSD License apply with one
@@ -14,6 +14,8 @@
 *
 * BSD License: http://www.opensource.org/licenses/bsd-license.php
 **/
+
+use PHPExperts\UserDirectory\Managers\UserManager;
 
 echo '<?xml version="1.0" encoding="UTF-8"?>' . "\n";
 ?>
